requirements.txt
django-heroku==0.3.1
gunicorn==20.0.4
dj-database-url==0.5.0
Django>=2.2.5
psycopg2==2.8.4
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
djangorestframework==3.11.1